TLDetector: Refactor the DummyDetector
