+++ ros/launch/site.launch
@@ -15,6 +15,10 @@
     <!--Traffic Light Locations and Camera Config -->
     <param name="traffic_light_config" textfile="$(find tl_detector)/site_traffic_light_config.yaml" />
 
+    <!-- Traffic Light Detector Params -->
+    <param name="traffic_light_detector" type="str" value="dummy" />
+    <!-- <param name="traffic_light_detector" type="str" value="dnn" /> -->
+
     <!--Traffic Light Detector Node -->
     <include file="$(find tl_detector)/launch/tl_detector_site.launch"/>
